@@ -31,13 +33,12 @@ versionProps.load(new FileInputStream(versionPropsFile))
 def verCode = versionProps['VERSION_CODE'].toInteger()
 
 android {
-    compileSdkVersion 29
-    buildToolsVersion "29.0.3"
+    compileSdkVersion 34
 
     defaultConfig {
         applicationId "com.guralnya.notification_tracker"
         minSdkVersion 23
-        targetSdkVersion 29
+        targetSdkVersion 34
         versionCode verCode
         versionName "1.1.$verCode"
 
@@ -88,9 +89,10 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
     }
+    namespace 'com.guralnya.notification_tracker'
 }
 
 def static getTime() {
